+++ Controller/Payment/Iframe/Loader.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * PayZen V2-Payment Module version 2.3.1 for Magento 2.x. Support contact : support@payzen.eu.
+ * PayZen V2-Payment Module version 2.3.2 for Magento 2.x. Support contact : support@payzen.eu.
  *
  * NOTICE OF LICENSE
  *
@@ -9,11 +9,11 @@
  * It is also available through the world-wide-web at this URL:
  * https://opensource.org/licenses/osl-3.0.php
  *
+ * @category  Payment
+ * @package   Payzen
  * @author    Lyra Network (http://www.lyra-network.com/)
  * @copyright 2014-2018 Lyra Network and contributors
  * @license   https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- * @category  payment
- * @package   payzen
  */
 namespace Lyranetwork\Payzen\Controller\Payment\Iframe;
 
@@ -48,7 +48,7 @@ class Loader extends \Magento\Framework\App\Action\Action
         $assets = $resultPage->getConfig()
             ->getAssetCollection()
             ->getAll();
-        foreach ($assets as $identifier => $asset) {
+        foreach (array_keys($assets) as $identifier) {
             $resultPage->getConfig()
                 ->getAssetCollection()
                 ->remove($identifier);