+++ Model/Api/PayzenResponse.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * PayZen V2-Payment Module version 2.3.1 for Magento 2.x. Support contact : support@payzen.eu.
+ * PayZen V2-Payment Module version 2.3.2 for Magento 2.x. Support contact : support@payzen.eu.
  *
  * NOTICE OF LICENSE
  *
@@ -9,11 +9,11 @@
  * It is also available through the world-wide-web at this URL:
  * https://opensource.org/licenses/osl-3.0.php
  *
+ * @category  Payment
+ * @package   Payzen
  * @author    Lyra Network (http://www.lyra-network.com/)
  * @copyright 2014-2018 Lyra Network and contributors
  * @license   https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- * @category  payment
- * @package   payzen
  */
 namespace Lyranetwork\Payzen\Model\Api;
 
@@ -145,7 +145,8 @@ if (! class_exists('PayzenResponse', false)) {
                 'AUTHORISED',
                 'AUTHORISED_TO_VALIDATE',
                 'CAPTURED',
-                'CAPTURE_FAILED' /* capture will be redone */
+                'CAPTURE_FAILED', /* capture will be redone */
+                'ACCEPTED'
             );
 
             return in_array($this->transStatus, $confirmedStatuses) || $this->isPendingPayment();
@@ -162,7 +163,8 @@ if (! class_exists('PayzenResponse', false)) {
                 'INITIAL',
                 'WAITING_AUTHORISATION',
                 'WAITING_AUTHORISATION_TO_VALIDATE',
-                'UNDER_VERIFICATION'
+                'UNDER_VERIFICATION',
+                'WAITING_FOR_PAYMENT'
             );
 
             return in_array($this->transStatus, $pendingStatuses);
@@ -472,11 +474,11 @@ if (! class_exists('PayzenResponse', false)) {
          * @param boolean $appendCode
          * @return string
          */
-        public static function translate($result, $result_type = self::TYPE_RESULT, $lang = 'fr', $appendCode = false)
+        public static function translate($result, $result_type = self::TYPE_RESULT, $lang = 'en', $appendCode = false)
         {
             // if language is not supported, use the domain default language
             if (!key_exists($lang, self::$RESPONSE_TRANS)) {
-                $lang = 'fr';
+                $lang = 'en';
             }
 
             $translations = self::$RESPONSE_TRANS[$lang];